changed .ex package to .util and refactored
