+++ src/main/java/com/amihaiemil/eva/util/BinaryArraySolution.java
@@ -25,7 +25,7 @@
  OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
  THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.amihaiemil.eva.ex;
+package com.amihaiemil.eva.util;
 
 import com.amihaiemil.eva.Solution;
