+++ src/test/java/com/amihaiemil/eva/EvolutionaryAlgorithmTest.java
@@ -27,9 +27,8 @@
  */
 package com.amihaiemil.eva;
 
-import com.amihaiemil.eva.ex.BinaryArraySolutionsGenerator;
-import com.amihaiemil.eva.ex.FitnessForBackpack;
-import com.amihaiemil.eva.ex.FitnessForBackpackEvaluator;
+import com.amihaiemil.eva.util.BinaryArraySolutionsGenerator;
+import com.amihaiemil.eva.util.FitnessForBackpackEvaluator;
 import org.junit.Test;
 
 import java.util.ArrayList;
@@ -52,12 +51,12 @@ public class EvolutionaryAlgorithmTest {
      */
     @Test
     public void simpleEvaFindsASolution() throws Exception {
-        List<Integer> weights = new ArrayList<Integer>();
+        long[] weights = new long[10];
         for(int i=0;i<10;i++) {
-            weights.add(r.nextInt(100));
+            weights[i] = r.nextInt(100);
         }
-        BinaryArraySolutionsGenerator generator = new BinaryArraySolutionsGenerator(weights.size());
         FitnessForBackpackEvaluator evaluator = new FitnessForBackpackEvaluator(weights);
+        BinaryArraySolutionsGenerator generator = new BinaryArraySolutionsGenerator(weights.length);
 
          // The algorithm should stop only when the found solution is acceptable.
         Condition solutionIsAcceptable = new Condition() {
@@ -68,10 +67,9 @@ public class EvolutionaryAlgorithmTest {
 
         Eva algorithm = new SimpleEvolutionaryAlgorithm();
         Solution solution = algorithm.with(generator).with(evaluator).with(solutionIsAcceptable).calculate();
-        int solutionWeight = ((FitnessForBackpack) solution.getFitness()).getWeight();
         assertTrue(
-                "Expected solution with weight <= 100, got weight: " + solutionWeight,
-                solutionWeight <= 100
+            "Expected solution with weight <= 100",
+            solutionIsAcceptable.passed(solution)
         );
     }
 }
